requirements.txt
numpy
pandas
python-dateutil
scikit-learn
plotly
matplotlib
